ovn: Do not expect the separate plugin package
OVN is part of Neutron since Ussuri [1]. As such, there is no longer a
separate python-networking-ovn package that needs to be installed.
This change removes the logic to install the separate plugin package,
and also updates the config file path to use the file provided by
the current neutron package.

In addition, this change ensures the ovn-metadata-agent package is
installed in Ubuntu/Debian as well.

[1] https://docs.openstack.org/releasenotes/neutron/ussuri.html#relnotes-16-0-0-stable-ussuri

Closes-Bug: #1945973
Change-Id: I8e32e0ca02b841058dc71f00a72b825e6441c1d3