Clean up direct dependencies on puppetlabs-apache
This change removes direct reference to some classes in
puppetlabs-apache. Details are explained below.

- The api class doesn't need access to anything defined in
  apache::params

- The following classes are included by the openstacklib::wsgi::apache
  resource type, and current inclusions are just redundant.

Dependencies defined in neutron::wsgi::apache are removed and loosen
to avoid direct deference to resources from puppetlabs-apache. When
httpd is used to run neutron-api, the neutron-service tag is added to
the httpd service, so that it is started during service phase.

Change-Id: I1a17a5ac990dd8e0a166b7d78b8a58a4b1055502