Accept system scope credential for Ironic API request
Currently Neutron uses the user credential in [ironic] section to use
External Event API in Ironic but this API is available only for system
admin when SRBAC is enforced.

Also, networking-baremetal uses the user credentials in [ironic]
section to list all ports, which is available only for system reader
when SRBAC is enforced.

This change allows usage of system-scoped credential instead of
project-scoped one for these two features.

Change-Id: I3bf7b285383ab4494cfa21440d9dd3fe26dd23df