Add support for [nova] live_migation_events
This change introduces capability to manage [nova] live_migation_events
which was introduced recently. This is implemented mainly so that we
can backport this feature to stable branches. The parameter is already
deprecated in master so a separate change to deprecate the parameter
will be submitted.

Conflicts:
	manifests/server/notifications/nova.pp

Change-Id: Ieccaf6b81939119c51ec74c44ee0fd2fcea3d99f
(cherry picked from commit 3389f02d67a14e10ffa5b988229507e9ce9998d0)