Deprecate support for networking-bigswitch
The networking-bigswitch plugin is unmaintained. The repo[1] has not
been updated for 2 years and no release has been created since Train.
Because we don't expect any user is using that unmainitained plugin
with recent versions of OpenStack, let's deprecate support for
the plugin so that we can remove it completely in the next cycle.

Closes-Bug: #1962579
Change-Id: Id4b07727af42b6a1d1145b71fc79f4bac7617eb3