Deprecate the mech_driver defined resource type
This defined type manages two configurations for SRIOV but both of
these are required for ancient releases of Neutron and are invalid in
recent releases.

1) The ml2_conf_sriov.ini file was removed during Ocata cycle by [1],
   and now no additional file is required by neutron-server.

2) The supported_pci_vendor_devs was deprecated during Newton cycle[2]
   and was removed during Ocata cycle[1].

[1] https://review.opendev.org/c/openstack/neutron/+/413439
[2] https://review.opendev.org/c/openstack/neutron/+/391911

This deprecates the mech_driver defined resource type because now it
has no actual logics. We've implemented separate classes for individual
mechanism driver and this common implementation is no longer required.

Change-Id: I4b3a6ecf627c6d7203d51f13d69a9400085f8df6