bgpvpn: Ensure the "default" service providers
The bgpvpn service plugin package provides its additional default value
in its configuration file[1]. This effectively appends the value to
Neutron's default and is required to add the service plugin properly.

Currently, the parameter is unmanaged when $::os_service_default is
used to keep that additional default value. However this implementation
leaves the old value if a user configures the parameter and later
removes it.

This change makes the class to ensure that additional default value,
to avoid such situation.

[1] https://opendev.org/openstack/networking-bgpvpn/src/branch/master/etc/neutron/networking_bgpvpn.conf

Change-Id: I5fef856aaf1f6e079e484197bd9d844577fc93d0