mlnx: Allow workaround for duplicate resources
The commit b2e7de79df98a986768aaef203ddbe02bee41072 , which was merged
during Train cycle, introduced a few new parameters to support multi
interface driver, but the following two parameters are already declared
in the corresponding agent classes and trigger duplicate resource
declaration when these classes are used in a single manifest.
 - dhcp_broadcast_reply
 - interface_driver

This change allows users to workaround the problem. Details are
documented in the release note.

Closes-Bug: #1987460
Change-Id: I993523dc7f0ec8ad90a50d049dc0f5ccf9f2b191