Fix wrong bridge name if bridge_mappings is commma-separated string
This change ensures the value is properly translated to a list even
if it's a flat string so that the bridge name is properly picked up.

This also fixes a few wrong assertions done in unit tests.

Closes-Bug: #2003324
Change-Id: I3a784588435619418cfd51a1ceffd9fd0ecd5e63
(cherry picked from commit 05b043698f1fc7511e3087bea6a4dada515e54a9)