Deprecate client class
The neutron CLI has been deprecated[1] and will be removed in a future
release. This deprecates the class to install the neutronclient package
because this class aims to help installing the package for CLI usage.
It should be installed by package dependencies in case it is used as
a library.

[1] 3a64a7a166be25d40436fd40c8351a79267bd3c4

Change-Id: I6ae441ea74c9e51e61a944583092bc0e0bc687f4