vpnaas: Do not install wrong backend package for OpenSwan driver
Currently the neutron::agents::vpnaas class installs LibreSwan (in
CentOS) or strongSwan (in Debian or Ubuntu) for the OpenSwan driver
but the OpenSwan driver only works with OpenSwan which is no longer
shipped. (That's why they developed different drivers for different
libraries).

This change drops installation of the wrong package and also deprecates
support for OpenSwan driver.

Change-Id: I158dd3411900241950ead9635bba4583eebcb983