Replace unnecessary ensure_resource
Some of the packages managed via ensure_resource are defined in only
the single class so we don't need ensure_resource.

Also titles of the package resources are updated so that users can make
references to the resources with a static name.

In addition, some resource notifications are redundant because the same
notification is implemented using anchors.

Change-Id: I0aeaca27d6a05c711a744a3ea95dda6fa7e2cbf3