Debian/Ubuntu: Prohibit BaGPipe agent
Currently networking-bagpipe is shipped in Debian and Ubuntu but their
packages do not provide the service unit for the agent (bagpipe-bgp).

Also the configuration path is different from the current expectation.

This makes the agent class hard-fail in case it's used in Debian or
Ubuntu, instead of making wrong configurations (or failures during
applying manifests). We'll revisit this if the package problem is
resolved and we hear a real request to support the service in Debian
and Ubuntu.

Closes-Bug: #2043283
Change-Id: If53fe4c586670e25b3dad6e712f5d7fb5778c341